+++ workspaces/marketplace/plugins/catalog-backend-module-marketplace/src/module.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright Red Hat, Inc.
+ * Copyright The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,9 +21,9 @@ import {
 import { catalogProcessingExtensionPoint } from '@backstage/plugin-catalog-node/alpha';
 
 import { MarketplacePluginProcessor } from './processors/MarketplacePluginProcessor';
-import { MarketplacePluginListProcessor } from './processors/MarketplacePluginListProcessor';
-import { DynamicPluginInstallStatusProcessor } from './processors/DynamicPluginInstallStatusProcessor';
-import { LocalPluginInstallStatusProcessor } from './processors/LocalPluginInstallStatusProcessor';
+import { MarketplaceCollectionProcessor } from './processors/MarketplaceCollectionProcessor';
+import { DynamicPackageInstallStatusProcessor } from './processors/DynamicPackageInstallStatusProcessor';
+import { LocalPackageInstallStatusProcessor } from './processors/LocalPackageInstallStatusProcessor';
 import { MarketplacePackageProcessor } from './processors/MarketplacePackageProcessor';
 
 /**
@@ -43,11 +43,11 @@ export const catalogModuleMarketplace = createBackendModule({
       async init({ logger, catalog, discovery, auth }) {
         logger.info('Adding Marketplace processors to catalog...');
         catalog.addProcessor(new MarketplacePluginProcessor());
-        catalog.addProcessor(new MarketplacePluginListProcessor());
-        catalog.addProcessor(new LocalPluginInstallStatusProcessor());
+        catalog.addProcessor(new MarketplaceCollectionProcessor());
+        catalog.addProcessor(new LocalPackageInstallStatusProcessor());
         catalog.addProcessor(new MarketplacePackageProcessor());
         catalog.addProcessor(
-          new DynamicPluginInstallStatusProcessor(discovery, auth),
+          new DynamicPackageInstallStatusProcessor(discovery, auth),
         );
       },
     });