+++ workspaces/marketplace/plugins/marketplace-common/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @red-hat-developer-hub/backstage-plugin-marketplace-common
 
+## 0.1.2
+
+### Patch Changes
+
+- a7e38a9: remove core-plugin-api dependency from common
+
+## 0.1.1
+
+### Patch Changes
+
+- fa70a22: remove unused knex and zod dependency from common package
+
+## 0.1.0
+
+### Minor Changes
+
+- d39d4e3: Initial tech-preview of the marketplace
+
+### Patch Changes
+
+- f16dd69: Refactore and update API, UI, and catalog types
+
 ## 0.0.9
 
 ### Patch Changes